Implemented Start controller with Game Manager, please note there are errors when running start game as the setup game controller has not been implemented

+++ src/main/java/seng201/team0/models/Towers/Farm.java
@@ -7,7 +7,7 @@ public class Farm extends Tower{
     private static final Resource Food = new Food();
     public static String towerImagePath = "resources/images/towers/farm_tower.png";
     public static String brokenImagePath = "resources/images/towers/farm_tower_broken.png";
-    public Farm(int inputReloadSpeed, double inputCost) {
+    public Farm(int inputReloadSpeed, int inputCost) {
         super(Food, inputReloadSpeed, Food.getResourceType(), "Resource: " + Food.getResourceType(), inputCost, towerImagePath, brokenImagePath);
     }
 }
