+++ src/main/java/seng201/team0/models/Towers/Sawmill.java
@@ -7,7 +7,7 @@ public class Sawmill extends Tower{
     private static final Resource Wood = new Wood();
     public static String towerImagePath = "resources/images/towers/sawmill_tower.png";
     public static String brokenImagePath = "resources/images/towers/sawmill_tower_broken.png";
-    public Sawmill(int inputReloadSpeed, double inputCost) {
+    public Sawmill(int inputReloadSpeed, int inputCost) {
         super(Wood, inputReloadSpeed, Wood.getResourceType(), "Resource: " + Wood.getResourceType(), inputCost, towerImagePath, brokenImagePath);
     }
 }
