+++ src/main/java/seng201/team0/models/Upgrades/Upgrade.java
@@ -5,7 +5,7 @@ import seng201.team0.models.Towers.Tower;
 
 public abstract class Upgrade extends Item {
     int upgradeCount;
-    public Upgrade(String inputName, String inputDescription, double inputCost) {
+    public Upgrade(String inputName, String inputDescription, int inputCost) {
         super(inputName, inputDescription, inputCost);
     }
     public void increaseCount() {
